Ran code cleanup

+++ com.revolsys.open.core/src/test/java/com/revolsys/jts/testold/algorithm/CGAlgorithmsTest.java
@@ -19,30 +19,30 @@ public class CGAlgorithmsTest extends TestCase {
 
   public void testDistanceLineLineDisjointCollinear() {
     assertEquals(1.999699, LineSegmentUtil.distanceLineLine(new PointDouble(
-    (double)0.0, 0), new PointDouble((double)9.9, 1.4), new PointDouble((double)11.88, 1.68), new PointDouble(
-      (double)21.78, 3.08)), 0.000001);
+      0.0, 0), new PointDouble(9.9, 1.4), new PointDouble(11.88, 1.68), new PointDouble(
+        21.78, 3.08)), 0.000001);
   }
 
   public void testDistancePointLine() {
-    assertEquals(0.5, LineSegmentUtil.distanceLinePoint(new PointDouble((double)0.0, 0), new PointDouble((double)1.0, 0), new PointDouble(
-    (double)0.5, 0.5)), 0.000001);
-    assertEquals(1.0, LineSegmentUtil.distanceLinePoint(new PointDouble((double)0.0, 0), new PointDouble((double)1.0, 0), new PointDouble(
-    (double)2.0, 0)), 0.000001);
+    assertEquals(0.5, LineSegmentUtil.distanceLinePoint(new PointDouble(0.0, 0), new PointDouble(1.0, 0), new PointDouble(
+      0.5, 0.5)), 0.000001);
+    assertEquals(1.0, LineSegmentUtil.distanceLinePoint(new PointDouble(0.0, 0), new PointDouble(1.0, 0), new PointDouble(
+      2.0, 0)), 0.000001);
   }
 
   public void testDistancePointLinePerpendicular() {
     assertEquals(0.5, CGAlgorithms.distancePointLinePerpendicular(
-      new PointDouble((double)0.5, 0.5), new PointDouble(
-        (double)0.0, 0), new PointDouble((double)1.0, 0,
-        Point.NULL_ORDINATE)), 0.000001);
+      new PointDouble(0.5, 0.5), new PointDouble(
+        0.0, 0), new PointDouble(1.0, 0,
+          Point.NULL_ORDINATE)), 0.000001);
     assertEquals(0.5, CGAlgorithms.distancePointLinePerpendicular(
-      new PointDouble((double)3.5, 0.5), new PointDouble(
-        (double)0.0, 0), new PointDouble((double)1.0, 0,
-        Point.NULL_ORDINATE)), 0.000001);
+      new PointDouble(3.5, 0.5), new PointDouble(
+        0.0, 0), new PointDouble(1.0, 0,
+          Point.NULL_ORDINATE)), 0.000001);
     assertEquals(0.707106, CGAlgorithms.distancePointLinePerpendicular(
-      new PointDouble((double)1.0, 0), new PointDouble((double)0.0,
-        0, Point.NULL_ORDINATE), new PointDouble((double)1.0, 1,
-        Point.NULL_ORDINATE)), 0.000001);
+      new PointDouble(1.0, 0), new PointDouble(0.0,
+        0, Point.NULL_ORDINATE), new PointDouble(1.0, 1,
+          Point.NULL_ORDINATE)), 0.000001);
   }
 
 }
