+++ com.revolsys.open.core/src/test/java/com/revolsys/jts/testold/geom/CoordinateArraysTest.java
@@ -59,26 +59,26 @@ public class CoordinateArraysTest extends TestCase {
   public void testPtNotInList1() {
     assertTrue(CoordinateArrays.ptNotInList(
       new Point[] {
-        new PointDouble((double)1.0, 1, Point.NULL_ORDINATE),
-        new PointDouble((double)2.0, 2, Point.NULL_ORDINATE),
-        new PointDouble((double)3.0, 3, Point.NULL_ORDINATE)
+        new PointDouble(1.0, 1, Point.NULL_ORDINATE),
+        new PointDouble(2.0, 2, Point.NULL_ORDINATE),
+        new PointDouble(3.0, 3, Point.NULL_ORDINATE)
       },
       new Point[] {
-        new PointDouble((double)1.0, 1, Point.NULL_ORDINATE),
-        new PointDouble((double)1.0, 2, Point.NULL_ORDINATE),
-        new PointDouble((double)1.0, 3, Point.NULL_ORDINATE)
+        new PointDouble(1.0, 1, Point.NULL_ORDINATE),
+        new PointDouble(1.0, 2, Point.NULL_ORDINATE),
+        new PointDouble(1.0, 3, Point.NULL_ORDINATE)
       }).equals(2,new PointDouble((double)2, 2, Point.NULL_ORDINATE)));
   }
 
   public void testPtNotInList2() {
     assertTrue(CoordinateArrays.ptNotInList(new Point[] {
-      new PointDouble((double)1.0, 1, Point.NULL_ORDINATE),
-      new PointDouble((double)2.0, 2, Point.NULL_ORDINATE),
-      new PointDouble((double)3.0, 3, Point.NULL_ORDINATE)
+      new PointDouble(1.0, 1, Point.NULL_ORDINATE),
+      new PointDouble(2.0, 2, Point.NULL_ORDINATE),
+      new PointDouble(3.0, 3, Point.NULL_ORDINATE)
     }, new Point[] {
-      new PointDouble((double)1.0, 1, Point.NULL_ORDINATE),
-      new PointDouble((double)2.0, 2, Point.NULL_ORDINATE),
-      new PointDouble((double)3.0, 3, Point.NULL_ORDINATE)
+      new PointDouble(1.0, 1, Point.NULL_ORDINATE),
+      new PointDouble(2.0, 2, Point.NULL_ORDINATE),
+      new PointDouble(3.0, 3, Point.NULL_ORDINATE)
     }) == null);
   }
 }
