+++ com.revolsys.open.core/src/test/java/com/revolsys/jts/testold/noding/SegmentPointComparatorTest.java
@@ -24,8 +24,7 @@ public class SegmentPointComparatorTest extends TestCase {
   private void checkNodePosition(final int octant, final double x0,
     final double y0, final double x1, final double y1,
     final int expectedPositionValue) {
-    final int posValue = SegmentPointComparator.compare(octant, new PointDouble((double)
-      x0, y0, Point.NULL_ORDINATE), new PointDouble((double)x1, y1, Point.NULL_ORDINATE));
+    final int posValue = SegmentPointComparator.compare(octant, new PointDouble(x0, y0, Point.NULL_ORDINATE), new PointDouble(x1, y1, Point.NULL_ORDINATE));
     assertTrue(posValue == expectedPositionValue);
   }
